Fix a couple of issues when downloading posts
Namely, make the code more tolerant to HTTP failures, such as when a domain is no longer found.

Add the ability to skip a post if it's a Comment or a self post, those don't have Images at all

Fix also an issue with a loop in which, if a listing has less than 25 posts and is the last page, the next_page ID will stay the same, so the code was attempting to download the same posts in the final listing over and over again until the page limit was hit. The code now recognizes when the last page ID is the same as the next page ID.